move class to right package
